fix: Move core to be in peerDependencies
## Motivation

Core package is used as singleton and when having more than single instance of core app config will stop working. While this is only temporary fix for the invalid architecture it will provide cleaner way to manage core<> app dependency.

How things will work now:
- App will be required to configure entire platform and setup metrics
- Each individual package will be relying on the core provided in app dependency.

Eventually we can map app package to core and let it be used inside every SDK - however this will affect documentation and end user API so not sure if we are ok with this move