Add ability to pass update functions to the restore handler (#292)
The first iteration heavily inspired by https://medium.com/twostoryrobot/a-recipe-for-offline-support-in-react-apollo-571ad7e6f7f4

I haven't resolved "chicken or egg problem" related to the fact that the client is usually created first and then update functions follow. This will mean that our update functions need to be pulled into a separate class (later we can replace them with the helper that generates both optimistic responses and also update function. That will greatly improve the user experience.  

EDIT: It is actually easy to describe that now and it has very little impact on users. I will need to create follow up ticket to document this pattern.

## Tasks
- [x] Refactor previous implementation
- [x] Integrate with a showcase
- [x] Add case to integration tests