fix: use optimistic data if available. add client to resolve.
we need to use the parent optimistic data when available as this is most
likely what the user will be looking at when interacting with an object.
also, the client should be added to the resolve function for the case
where client diff is empty because of an optimistic response.